refactor: 패키지명 수정
기존: business
변경: application
