+++ src/main/java/codesquad/application/handler/LoginHandler.java
@@ -1,7 +1,7 @@
-package codesquad.business.handler;
+package codesquad.application.handler;
 
-import codesquad.business.model.User;
-import codesquad.business.persistence.UserRepository;
+import codesquad.application.model.User;
+import codesquad.application.persistence.UserRepository;
 import codesquad.http.cookie.Cookie;
 import codesquad.http.handler.AbstractDynamicRequestHandler;
 import codesquad.http.message.HttpRequest;
