+++ src/test/java/codesquad/application/handler/LoginHandlerTest.java
@@ -1,9 +1,9 @@
-package codesquad.business.handler;
+package codesquad.application.handler;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import codesquad.business.model.User;
-import codesquad.business.persistence.UserRepository;
+import codesquad.application.model.User;
+import codesquad.application.persistence.UserRepository;
 import codesquad.environment.TestEnvironment;
 import codesquad.http.cookie.Cookie;
 import codesquad.http.message.HttpRequest;
